登録の処理のための API 通信中にスピナーを表示し、操作ができないようにしました。
操作が渋滞するとうまく動作しない危険性があるからです。
また、データを取ってくるのに時間がかかり、画面が動かないときに
処理中であることをユーザーに分かりやすく示す役割もあります。
